refactor(admin-v2): extract lead-profile helpers + split components (P3-3)
AdminV2LeadProfile.jsx was 1,966 lines hosting ~21 sub-components alongside the
dense pure functions that decide what the page actually says. Split three ways,
verbatim:

  src/lib/adminV2/leadHistory.js          the pure data layer
  src/pages/adminv2/leadProfile/          the Lead Score card + shared primitives
  AdminV2LeadProfile.jsx                  the page itself, 1,966 → 976 lines

buildHistory, heroFor and deliveryState were the point. They were
module-private, so the only way to exercise any of them was to mount the whole
page with a full profile payload — which nothing did. They decide what an
operator reads about a real person's lead and they were effectively untested.
They are now importable and directly tested (MarketplaceFlow already exports its
helpers "for tests"; this follows that).

39 new cases cover the parts where being wrong is expensive and silent: that an
accepted WhatsApp message is never reported as delivered, that Meta error 131049
is named as the per-person marketing cap rather than a generic failure an
operator would pointlessly retry, that a draw lapse is timelined at the lapse and
not at the draw, that named assignments suppress the raw uuid duplicates, and
that a pre-2026-07-10 signup reads "unrecorded" rather than "unverified".

Verified non-vacuous: three mutations to the moved logic (disabling the 131049
branch, dropping the NaN-timestamp guard, reversing the sort) fail 5 of them.

Nothing visual changed. Existing AdminV2LeadProfile tests pass unchanged; the
whole frontend suite is 145 files / 1,847 tests green, eslint clean, and both
brand builds succeed.
