+++ src/pages/adminv2/AdminV2LeadProfile.jsx
@@ -20,12 +20,11 @@ import { Link, useLocation, useNavigate, useParams, useSearchParams } from 'reac
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 import { useProspectProfile, useAgentOptions } from '@/hooks/queries/useAdminV2';
-import { bulkAssign, bulkReturnToHeld, bulkDelete, fetchConsentCopy, fetchScoringEdition } from '@/api/adminV2';
+import { bulkAssign, bulkReturnToHeld, bulkDelete, fetchConsentCopy } from '@/api/adminV2';
 import { STATUS_LABELS, SOURCE_LABELS, heldLabel } from '@/lib/adminV2/constants';
 import { fmtDateTime, fmtDate, fmtDay, fmtSGDExact } from '@/lib/adminV2/format';
 import { Card, Chip, Skeleton, ErrorState, EmptyState } from '@/components/adminv2/primitives';
-import { COMPONENT_LABELS } from '@/lib/adminV2/scoringLabels';
-import { rowChipFor, drawWindowDay } from '@/lib/adminV2/outcome';
+import { rowChipFor } from '@/lib/adminV2/outcome';
 import {
   AlertDialog, AlertDialogContent, AlertDialogHeader, AlertDialogTitle,
   AlertDialogDescription, AlertDialogFooter, AlertDialogCancel, AlertDialogAction,
@@ -33,44 +32,27 @@ import {
 import { useAdminV2Mobile } from '@/components/adminv2/mobile/useAdminV2Mobile';
 import MobileSheet, { SheetHead, SheetMenuItem } from '@/components/adminv2/mobile/MobileSheet';
 import { MobileActionBar, MobilePrimaryBtn, MobileMoreBtn } from '@/components/adminv2/mobile/MobileBars';
+// The page's pure data layer — extracted so it can be tested directly (P3-3).
+import {
+  fullName,
+  sameName,
+  sgtTime,
+  sgtDayKey,
+  fmtPhone,
+  campaignAccent,
+  evidenceOf,
+  heroFor,
+  receiptBits,
+  buildHistory,
+} from '@/lib/adminV2/leadHistory';
+// The page's shared presentational primitives and the Lead Score card, split
+// out of this file (P3-3) — same markup, just no longer interleaved.
+import { HoverHint, KvRow, Disclosure, GlyphTile } from './leadProfile/primitives';
+import { LeadScoreCard, LeadScoreBadge, EnrichmentCard } from './leadProfile/LeadScoreCard';
 
 // ── shared bits ─────────────────────────────────────────────────────────────
 
-const SGT = 'Asia/Singapore';
-const fullName = (o) => `${o?.firstName || ''} ${o?.lastName || ''}`.trim();
-const sameName = (a, b) => fullName(a).toLowerCase() === fullName(b).toLowerCase();
-// hourCycle 'h23', not hour12:false — ICU versions map the latter to h24 or
-// h23 depending on runtime, so midnight was "24:xx" on some (fmtDateTime has
-// the full story).
-const sgtTime = (v) => new Date(v).toLocaleTimeString('en-SG', { hour: '2-digit', minute: '2-digit', hourCycle: 'h23', timeZone: SGT });
-const sgtDayKey = (v) => new Intl.DateTimeFormat('en-CA', { timeZone: SGT }).format(new Date(v));
-const drawWindowDayUpper = (v) => drawWindowDay(v).toUpperCase();
-/** House phone display: +6591234567 → "+65 9123 4567" (root CLAUDE.md rule). */
-const fmtPhone = (v) => {
-  const m = /^\+65(\d{4})(\d{4})$/.exec(String(v || ''));
-  return m ? `+65 ${m[1]} ${m[2]}` : (v || null);
-};
-/** History rows carry a per-campaign colored dot (hover = full campaign name)
- * instead of a truncated text tag — "IPHONE" couldn't tell two iPhone
- * campaigns apart and read as a device name. Same palette + hash as the
- * Redemptions console's campaign accents, keyed by campaign id so twin-named
- * campaigns still get distinct colors. */
-const CAMPAIGN_ACCENTS = ['#0364D3', '#6A3FD1', '#8F6400', '#177239', '#BD3A2E', '#0E7490'];
-function campaignAccent(key) {
-  let hash = 0;
-  const s = String(key || '');
-  for (let i = 0; i < s.length; i += 1) hash = (hash * 31 + s.charCodeAt(i)) >>> 0;
-  return CAMPAIGN_ACCENTS[hash % CAMPAIGN_ACCENTS.length];
-}
-const campaignRef = (id, name) => (name || id ? { key: String(id || name), name: name || 'Unnamed campaign' } : null);
 
-const BOOST_VIA_COPY = { agent_scan: 'consultant scan', agent_button: 'consultant confirmation' };
-const NOT_ELIGIBLE_COPY = {
-  no_phone: 'no phone on record',
-  phone_unverified: 'phone unverified',
-  terms_not_pinned: 'draw terms not accepted',
-  signed_up_after_close: 'signed up after entries closed',
-};
 /** Tri-state OTP evidence (backend `verificationEvidence`). "UNRECORDED" is
  * deliberately not "UNVERIFIED" — for signups before the 2026-07-09 stamp we
  * hold no proof either way, and the form has hard-gated submit on a passed OTP
@@ -81,183 +63,11 @@ const VERIFICATION_LABEL = {
   unrecorded: { text: 'UNRECORDED', hint: 'Signed up before the server stored OTP proof (2026-07-09). The form required a passed OTP to submit, so this is a gap in our records — not a lead that skipped verification.' },
 };
 
-/** Client-side twin of the backend's `phoneVerificationEvidence` — used only
- * for the no-linked-person fallback row, which is composed from the prospect
- * the page already holds rather than from a person payload. */
-const STAMP_EPOCH_MS = Date.parse('2026-07-10T00:00:00Z');
-function evidenceOf(prospect) {
-  if (prospect?.sourceMetadata?.phoneVerifiedAt) return 'verified';
-  const t = prospect?.createdAt ? Date.parse(prospect.createdAt) : NaN;
-  return Number.isFinite(t) && t < STAMP_EPOCH_MS ? 'unrecorded' : 'unverified';
-}
 
-const DIAGNOSTIC_COPY = {
-  no_active_activation: 'no reward attached to this campaign',
-  allocation_exhausted: 'quota full',
-  phone_not_verified: 'phone unverified',
-  no_phone: 'no phone on record',
-  duplicate_phone: 'this phone already holds one',
-  offer_not_active: 'offer paused',
-  activation_ended: 'activation ended',
-  quarantined: 'lead is held',
-  not_issued_yet: 'issuance sweep hasn’t landed',
-  // NOT "phone unverified": this signup predates the server-side OTP stamp
-  // (2026-07-09), and the public form has never let an unverified phone
-  // submit. The proof is missing, the verification probably wasn't.
-  verification_not_recorded: 'signed up before we recorded OTP proof',
-};
 
-/**
- * The outcome hero — the loudest element of a drill-in (24px/800). Returns
- * { big, tail, tone, meta } in the exact lifecycle-honest voices of the
- * backend contract; `tone` ∈ ink | ok | quiet.
- */
-function heroFor(draw, rewards, diagnostic) {
-  if (draw) {
-    switch (draw.state) {
-      case 'provisional_in':
-        return draw.boosted
-          ? { big: `On track for ×${draw.multiplier}`, tail: ` — ${BOOST_VIA_COPY[draw.boostVia] || 'boost'} recorded`, tone: 'ink', meta: draw.closesAt ? `CLOSES ${drawWindowDayUpper(draw.closesAt)}` : null }
-          : { big: '1 chance so far', tail: ' — boost window open', tone: 'ink', meta: (draw.boostClosesAt || draw.closesAt) ? `BOOST BY ${drawWindowDayUpper(draw.boostClosesAt || draw.closesAt)}` : null };
-      case 'provisional_out':
-        return { big: 'Not counted yet', tail: ` — ${NOT_ELIGIBLE_COPY[draw.notEligibleReason] || 'not eligible'}`, tone: 'quiet', meta: draw.closesAt ? `CLOSES ${drawWindowDayUpper(draw.closesAt)}` : null };
-      case 'frozen_in':
-        return { big: 'In the pool', tail: draw.boosted ? ` — ×${draw.multiplier} boost applies at seal` : ' — 1 chance', tone: 'ink', meta: 'FROZEN' };
-      case 'excluded_at_freeze':
-        return { big: 'Excluded at freeze', tail: '', tone: 'quiet', meta: null };
-      case 'sealed': {
-        const n = draw.chances;
-        const chances = `${n} chance${n === 1 ? '' : 's'}`;
-        const o = draw.outcome;
-        if (!o) return { big: chances, tail: ' — sealed', tone: 'ink', meta: 'SEALED' };
-        switch (o.status) {
-          case 'selected_pending': return { big: 'Selected', tail: ` — claim by ${fmtDate(o.claimDeadline)}`, tone: 'ink', meta: `ATTEMPT ${o.attemptNo}` };
-          case 'selected_claimed': return { big: '🏆 Winner', tail: ` — claimed ${fmtDate(o.claimedAt)}`, tone: 'ok', meta: null };
-          case 'not_selected_final': return { big: 'Not selected', tail: ` — ${chances}`, tone: 'quiet', meta: null };
-          case 'not_selected_yet': return { big: chances, tail: ' — draw in progress', tone: 'ink', meta: null };
-          default: return { big: `Selected — ${o.status.replace('selected_', '')}`, tail: ', redrawn', tone: 'quiet', meta: `ATTEMPT ${o.attemptNo}` };
-        }
-      }
-      case 'void': return { big: 'Draw void', tail: '', tone: 'quiet', meta: null };
-      case 'erased_draw_unavailable': return { big: '⊘ Draw record unavailable', tail: '', tone: 'quiet', meta: 'ERASED' };
-      case 'no_draw_record': return { big: 'Draw configured', tail: ' — no draw record yet', tone: 'quiet', meta: null };
-      default: break;
-    }
-  }
-  const ent = rewards?.[0];
-  if (ent) {
-    const title = ent.rewardTitle ? ` — ${ent.rewardTitle}` : '';
-    switch (ent.state) {
-      case 'reserved': return { big: 'Reserved', tail: title, tone: 'ink', meta: ent.expiresAt ? `EXPIRES ${fmtDate(ent.expiresAt).toUpperCase()}` : null };
-      case 'unlocked': return { big: 'Unlocked', tail: title, tone: 'ok', meta: ent.expiresAt ? `VOUCHER UNTIL ${fmtDate(ent.expiresAt).toUpperCase()}` : null };
-      case 'redeemed': return { big: 'Redeemed ✓', tail: title, tone: 'ok', meta: ent.redeemedAt ? fmtDate(ent.redeemedAt).toUpperCase() : null };
-      case 'expired': return { big: 'Expired', tail: title, tone: 'quiet', meta: null };
-      case 'blocked': return { big: 'Blocked', tail: title, tone: 'quiet', meta: null };
-      default: return { big: ent.state || ent.status, tail: title, tone: 'quiet', meta: null };
-    }
-  }
-  if (diagnostic) {
-    return diagnostic === 'not_issued_yet'
-      ? { big: 'Reward pending', tail: ` — ${DIAGNOSTIC_COPY[diagnostic]}`, tone: 'quiet', meta: null }
-      : { big: 'No reward', tail: ` — ${DIAGNOSTIC_COPY[diagnostic] || diagnostic}`, tone: 'quiet', meta: null };
-  }
-  return { big: 'No outcome recorded', tail: '', tone: 'quiet', meta: null };
-}
 
-// Post-acceptance truth (wa-delivery-truth): a receipt's base state means the
-// provider ACCEPTED the message — only the joined status inbox can prove
-// delivered/read, or surface Meta's silent drops (131049 = the per-user
-// marketing frequency cap that ate real boost receipts). Deliberately words,
-// not glyphs — a ✓ reads as "delivered" no matter what it technically meant —
-// and every state carries a hover explanation (native title).
-function deliveryState(r) {
-  if (!r) return null;
-  if (!r.ok) {
-    return {
-      ok: false,
-      text: 'failed to send',
-      hint: `Rejected before it left our system${r.error ? `: ${r.error}` : ''}.`,
-    };
-  }
-  const st = r.delivery?.status;
-  if (st === 'failed') {
-    const capped = String(r.delivery?.errorCode) === '131049';
-    return {
-      ok: false,
-      text: capped ? 'not delivered (Meta marketing limit)' : 'not delivered',
-      hint: capped
-        ? "Meta accepted it, then dropped it — this person hit Meta's per-person marketing message limit. Retries fail until the window clears; use email or a link."
-        : `Meta reported it undelivered${r.delivery?.errorTitle ? `: ${r.delivery.errorTitle}` : r.delivery?.errorCode ? ` (code ${r.delivery.errorCode})` : ''}.`,
-    };
-  }
-  if (st === 'read') {
-    return { ok: true, text: 'read', hint: 'Opened by the recipient on WhatsApp.' };
-  }
-  if (st === 'delivered') {
-    return { ok: true, text: 'delivered', hint: "Delivered to the recipient's device." };
-  }
-  if (st === 'sent') {
-    return { ok: true, text: 'sent, delivery pending', hint: 'Dispatched by Meta — delivery not confirmed yet.' };
-  }
-  return {
-    ok: true,
-    text: 'accepted',
-    hint: 'Accepted by the provider. Delivery is not confirmed for this message.',
-  };
-}
 
-/** Themed hover hint — dotted underline invites the hover; the bubble uses the
- * page's own surface tokens instead of the native OS tooltip. */
-function HoverHint({ label, hint, underline = true, color, ariaLabel }) {
-  const [open, setOpen] = useState(false);
-  return (
-    <span
-      aria-label={ariaLabel || undefined}
-      onMouseEnter={() => setOpen(true)}
-      onMouseLeave={() => setOpen(false)}
-      // Focusable because the hint is now the ONLY place several of these
-      // reasons exist — the scoring card moved them off the row. A hint you
-      // can only reach with a mouse is a hint a keyboard user cannot read at
-      // all. Escape closes it without moving focus, so a reader who opened one
-      // by tabbing is not trapped reading it.
-      tabIndex={hint ? 0 : undefined}
-      onFocus={() => setOpen(true)}
-      onBlur={() => setOpen(false)}
-      onKeyDown={(e) => { if (e.key === 'Escape' && open) { e.stopPropagation(); setOpen(false); } }}
-      style={{ position: 'relative', display: 'inline-block' }}
-    >
-      <span
-        style={{
-          ...(underline ? { textDecoration: 'underline dotted', textDecorationColor: 'var(--ink-3)', textUnderlineOffset: 3 } : {}),
-          ...(color ? { color } : {}),
-        }}
-      >{label}</span>
-      {open && hint && (
-        <span
-          role="tooltip"
-          style={{
-            position: 'absolute', left: 0, bottom: 'calc(100% + 7px)', zIndex: 40,
-            width: 230, padding: '8px 11px', background: 'var(--surface)',
-            border: '1px solid var(--line)', borderRadius: 9,
-            boxShadow: '0 10px 28px rgba(0,0,0,.3)',
-            fontSize: 11.5, lineHeight: 1.5, color: 'var(--ink-2)',
-            fontFamily: 'var(--font-ui)', fontWeight: 500,
-            whiteSpace: 'normal', textAlign: 'left', textDecoration: 'none',
-          }}
-        >{hint}</span>
-      )}
-    </span>
-  );
-}
 
-function receiptBits(delivery) {
-  if (!delivery) return [];
-  const bit = (r, prefix) => {
-    const s = deliveryState(r);
-    return s ? { ok: s.ok, prefix, state: s } : null;
-  };
-  return [bit(delivery.email, 'pass emailed'), bit(delivery.whatsapp, 'WhatsApp')].filter(Boolean);
-}
 
 /** The exact wording a consent version stamped — themed overlay, closes on
  * backdrop click or ✕. Draw-terms clauses arrive as pinned designer HTML
@@ -304,203 +114,6 @@ function ConsentCopyModal({ view, onClose }) {
 
 // ── history composition (family + tag per row; grouped by SGT day) ──────────
 
-const FAMILY_TILE = {
-  signup: { glyph: '●', bg: 'var(--accent-soft)', fg: 'var(--accent-text)' },
-  assignment: { glyph: '→', bg: 'var(--accent-soft)', fg: 'var(--accent-text)' },
-  unassignment: { glyph: '←', bg: 'var(--hold-soft)', fg: 'var(--hold)' },
-  boost: { glyph: '◆', bg: 'var(--hold-soft)', fg: 'var(--hold)' },
-  reward: { glyph: '◆', bg: 'var(--ok-soft)', fg: 'var(--ok)' },
-  outcome: { glyph: '★', bg: 'var(--ok-soft)', fg: 'var(--ok)' },
-  consent: { glyph: '✋', bg: 'var(--hold-soft)', fg: 'var(--hold)' },
-  screening: { glyph: '✦', bg: 'var(--ok-soft)', fg: 'var(--ok)' },
-  delivery: { glyph: '✉', bg: 'var(--surface-2)', fg: 'var(--ink-2)' },
-  arrival: { glyph: '○', bg: 'var(--surface-2)', fg: 'var(--ink-3)' },
-  generic: { glyph: '○', bg: 'var(--surface-2)', fg: 'var(--ink-2)' },
-};
-
-// Voucher/pass lifecycle rows (lead-history-completeness). `drawLinked`
-// decides the noun — event metadata alone can't (draw lookups fail open on
-// the writer side).
-const RESEND_NOUNS = { resend_pass: 'Pass', resend_voucher: 'Voucher', resend_boost: 'Boost receipt' };
-function lifecycleRow(ev, e) {
-  const noun = e.drawLinked ? 'Pass' : 'Voucher';
-  if (ev.type === 'verify_attempt') {
-    return { title: e.drawLinked ? 'Pass scanned by consultant' : 'Voucher scanned at merchant', quiet: true, family: 'reward' };
-  }
-  if (ev.type === 'rejected') {
-    return { title: 'Scan rejected — not redeemable', quiet: true, family: 'reward', tone: 'bad' };
-  }
-  if (ev.type === 'unlock_reversed') {
-    return { title: 'Draw boost undone', detail: ev.reason ? ` — ${ev.reason}` : null, family: 'unassignment' };
-  }
-  if (ev.type === 'reversed') {
-    return { title: `Redemption voided${ev.actorName ? ` by ${ev.actorName}` : ''}`, family: 'reward', tone: 'bad' };
-  }
-  if (ev.type === 'expired') {
-    // The sweep only expires ELIGIBLE rows — this event is reservation expiry.
-    return { title: e.drawLinked ? 'Draw pass expired' : 'Reservation expired', quiet: true, family: 'generic' };
-  }
-  if (ev.type === 'manual_override') {
-    if (ev.action === 'cancelled') {
-      return { title: `Cancelled${ev.actorName ? ` by ${ev.actorName}` : ''}`, detail: ev.reason ? ` — ${ev.reason}` : null, family: 'reward', tone: 'bad' };
-    }
-    if (ev.action === 'auto_resend') {
-      return { title: `${noun} resend retried automatically`, quiet: true, family: 'delivery' };
-    }
-    if (RESEND_NOUNS[ev.action]) {
-      const via = ev.channel === 'link' ? 'share link issued' : ev.channel ? `via ${ev.channel}` : '';
-      return {
-        // "initiated", deliberately — the audit row commits before the
-        // fire-and-forget send; the delivery receipt rows carry the outcome.
-        title: `${RESEND_NOUNS[ev.action]} resend initiated${ev.actorName ? ` by ${ev.actorName}` : ''}`,
-        detail: via ? ` — ${via}` : null,
-        quiet: true,
-        family: 'delivery',
-      };
-    }
-  }
-  return null;
-}
-
-const OUTCOME_LAPSE_COPY = {
-  selected_unclaimed: 'not claimed in time',
-  selected_unreachable: 'could not be reached',
-  selected_ineligible: 'ineligible',
-  selected_declined: 'declined the prize',
-};
-
-function buildHistory(p, journey) {
-  const events = [];
-  const push = (at, title, { detail = null, family = 'generic', quiet = false, campaign = null, state = null, tone = null } = {}) => {
-    const t = at ? Date.parse(at) : NaN;
-    if (!Number.isNaN(t)) events.push({ at: t, title, detail, family, quiet, campaign, state, tone });
-  };
-  const anchorCampaign = campaignRef(p.campaign?.id, p.campaign?.name);
-
-  // Journey signups carry NAMED assignment events (backend resolves the agent
-  // uuid); when present they replace the anchor's raw "Assigned to agent
-  // <uuid>" activity rows, which would otherwise duplicate them unreadably.
-  const journeyHasAssignments = (journey?.signups || []).some((s) => s.assignments?.length);
-  const rows = (Array.isArray(p.timeline)
-    ? p.timeline.map((x) => ({
-      at: x.row?.createdAt || x.row?.created_at,
-      label: x.entry?.label || x.row?.description || x.row?.type || 'activity',
-      type: x.row?.type || null,
-    }))
-    : (p.activities || []).map((a) => ({ at: a.createdAt, label: a.description || a.type, type: a.type || null }))
-  ).filter((r) => !(journeyHasAssignments && r.type === 'assigned'));
-  for (const r of rows) push(r.at, r.label, { quiet: true, campaign: anchorCampaign });
-
-  for (const s of journey?.signups || []) {
-    const campaign = campaignRef(s.campaign?.id, s.campaign?.name);
-    push(s.createdAt, `Signed up as ${fullName(s) || '—'}`, { detail: s.campaign?.name ? ` — ${s.campaign.name}` : null, family: 'signup', campaign });
-    for (const a of s.assignments || []) {
-      if (a.kind === 'returned_to_held') {
-        push(a.at, 'Returned to held', {
-          detail: a.agentName ? ` — taken back from ${a.agentName}` : null,
-          family: 'unassignment', campaign,
-        });
-      } else if (a.kind === 'unassigned') {
-        push(a.at, `Unassigned${a.agentName ? ` from ${a.agentName}` : ''}`, {
-          family: 'unassignment', campaign,
-        });
-      } else {
-        push(a.at, `Assigned to ${a.agentName || (a.external ? 'an MKTR Leads buyer' : 'an agent')}`, {
-          detail: a.external ? ' — external buyer' : null,
-          family: 'assignment', campaign,
-        });
-      }
-    }
-    if (s.draw?.boostedAt) {
-      push(s.draw.boostedAt, 'Draw boost recorded', { detail: ` — ${BOOST_VIA_COPY[s.draw.boostVia] || 'boost'}`, family: 'boost', campaign });
-    }
-    // Draw result (lead-history-completeness): selection at drawnAt, its
-    // resolution (claim / lapse) at outcomeAt — never timeline a lapse at the
-    // draw moment. Missing timestamps skip the row rather than invent one.
-    const oc = s.draw?.outcome;
-    if (oc?.status?.startsWith('selected_')) {
-      if (oc.drawnAt) {
-        push(oc.drawnAt, 'Selected in the draw', {
-          detail: oc.attemptNo > 1 ? ` — redraw ${oc.attemptNo}` : null,
-          family: 'outcome', campaign,
-        });
-      }
-      if (oc.status === 'selected_claimed' && (oc.claimedAt || oc.outcomeAt)) {
-        push(oc.claimedAt || oc.outcomeAt, 'Prize claimed', { family: 'outcome', campaign });
-      } else if (OUTCOME_LAPSE_COPY[oc.status] && oc.outcomeAt) {
-        push(oc.outcomeAt, 'Selection lapsed', { detail: ` — ${OUTCOME_LAPSE_COPY[oc.status]}`, family: 'outcome', tone: 'bad', campaign });
-      }
-    } else if (oc?.status === 'not_selected_final' && oc.outcomeAt) {
-      push(oc.outcomeAt, 'Not selected in the draw', { quiet: true, family: 'generic', campaign });
-    }
-  }
-  // Consumer-less (B4) leads carry their entitlements on signupProfile — the
-  // same lifecycle events must render for them too.
-  const entList = journey?.entitlements?.length ? journey.entitlements : (p.signupProfile?.entitlements || []);
-  for (const e of entList) {
-    const campaign = campaignRef(e.campaignId, e.campaignName);
-    const title = e.rewardTitle || 'reward';
-    push(e.createdAt, e.drawLinked ? 'Draw pass reserved' : 'Reward reserved', { detail: e.drawLinked ? null : ` — ${title}`, family: 'reward', quiet: true, campaign });
-    if (e.unlockedAt) push(e.unlockedAt, e.drawLinked ? 'Draw boost confirmed' : 'Voucher unlocked', { detail: e.drawLinked ? null : ` — ${title}`, family: e.drawLinked ? 'boost' : 'reward', campaign });
-    if (e.redeemedAt) push(e.redeemedAt, 'Voucher redeemed ✓', { detail: ` — ${title}`, family: 'reward', campaign });
-    if (e.claimViews?.firstAt) {
-      // First open only, with the total — a refresh-happy customer is one row.
-      push(e.claimViews.firstAt, `Opened their reward link${e.claimViews.count > 1 ? ` — ×${e.claimViews.count} views` : ''}`, { family: 'arrival', quiet: true, campaign });
-    }
-    for (const ev of e.events || []) {
-      const row = lifecycleRow(ev, e);
-      if (row) push(ev.at, row.title, { detail: row.detail || null, family: row.family, quiet: row.quiet || false, tone: row.tone || null, campaign });
-    }
-    // Issued vouchers expire WITHOUT a ledger event (the sweep only touches
-    // eligible rows) — derive the row from expiresAt.
-    if (!e.drawLinked && e.status === 'issued' && !e.redeemedAt && e.expiresAt && Date.parse(e.expiresAt) < Date.now()) {
-      push(e.expiresAt, 'Voucher expired', { detail: ` — ${title}`, family: 'generic', quiet: true, campaign });
-    }
-    for (const ch of ['email', 'whatsapp']) {
-      const r = e.delivery?.[ch];
-      if (!r?.at) continue;
-      push(r.at, `${(r.kind || 'pass').replace(/_/g, ' ')} ${ch === 'email' ? 'emailed' : 'WhatsApp'} — `, { family: 'delivery', quiet: true, campaign, state: deliveryState(r) });
-    }
-  }
-  // Consent lifecycle (contact kind, source-allowlisted server-side).
-  for (const c of journey?.consentTimeline || []) {
-    if (c.granted === false) {
-      const via = c.via === 'wa_stop' ? 'WhatsApp STOP' : c.via === 'unsubscribe_link' ? 'unsubscribe link' : (c.via || 'unsubscribe');
-      push(c.at, 'Marketing consent withdrawn', { detail: ` — ${via}`, family: 'consent', tone: 'bad' });
-    } else {
-      push(c.at, 'Marketing consent re-granted', { family: 'consent' });
-    }
-  }
-  if (journey?.consumer?.erasedAt) {
-    push(journey.consumer.erasedAt, 'Erased under PDPA', { detail: ' — personal data removed', family: 'unassignment', tone: 'bad' });
-  }
-  for (const b of journey?.broadcasts?.recent || []) {
-    push(b.sentAt || b.at, `Marketing email — ${b.status}${b.reason ? ` (${b.reason.replace(/_/g, ' ')})` : ''}`, { family: 'delivery', quiet: true });
-  }
-  for (const ld of p.lyfeDelivery || []) {
-    push(ld.at, `Lyfe ${ld.eventType.replace('lead.', '')} — ${ld.status}${ld.reason ? ` (${ld.reason.replace(/_/g, ' ')})` : ''}`, { family: 'delivery', quiet: true, campaign: anchorCampaign });
-  }
-  if (p.session?.startedAt) {
-    push(p.session.startedAt, `Arrived — ${p.session.landingPath || 'landing page'}`, { family: 'arrival', quiet: true, campaign: anchorCampaign });
-  }
-  for (const a of Object.values(p.screeningMetadata?.attempts || {})) {
-    if (!a?.startedAt) continue;
-    const outcome = (a.outcome || 'attempted').replace(/_/g, ' ');
-    push(a.startedAt, `Screening call — ${outcome}`, { family: 'screening', quiet: a.outcome !== 'qualified', campaign: anchorCampaign });
-  }
-  return events.sort((a, b) => b.at - a.at);
-}
-
-// ── small presentational pieces ─────────────────────────────────────────────
-
-function KvRow({ label, mono = true, children }) {
-  return (
-    <div style={{ display: 'flex', gap: 12, fontSize: 12.5, padding: '4px 0', alignItems: 'baseline' }}>
-      <span style={{ width: 96, flex: 'none', color: 'var(--ink-3)' }}>{label}</span>
-      <span style={{ fontFamily: mono ? 'var(--font-mono)' : 'var(--font-ui)', color: 'var(--ink)', minWidth: 0, overflowWrap: 'anywhere' }}>{children}</span>
-    </div>
-  );
-}
 
 function ConsentKv({ label, state, legacy }) {
   const granted = state ? state.granted : legacy;
@@ -518,607 +131,7 @@ function ConsentKv({ label, state, legacy }) {
   );
 }
 
-/* ── MEET × BUY scoring (consumer-profile-enrichment §7.1b, §8) ───────────
- * The BREAKDOWN LEADS and the number follows (plan §11 step 4): these weights
- * are a v1 calibration, so the panel has to show its working or it is just an
- * authoritative-looking guess. Every component states whether it was assessed
- * or is simply unknown — that distinction is the point, because "unknown
- * capacity" and "low capacity" must never render the same way.
- */
-// COMPONENT_LABELS moved to @/lib/adminV2/scoringLabels — shared with the
-// sheet editor (campaign-scoring-editor §3.1) so the card and the editor
-// describe the same components in the same words.
 
-const FACT_LABELS = {
-  'identity.gender': 'gender',
-  'identity.birth_year_band': 'born',
-  'identity.ethnicity': 'ethnicity',
-  'identity.preferred_language': 'language',
-  'family.marital_status': 'marital status',
-  'family.children': 'children',
-  'family.children_count_band': 'children',
-  'family.parents_alive': 'parents alive',
-  'household.pets': 'pets',
-  'assets.car_owner': 'car owner',
-  'assets.property': 'property',
-  'career.job_title': 'job title',
-  'career.industry': 'industry',
-  'career.employment': 'employment',
-  'finance.income_band': 'monthly income',
-  'finance.annual_income_band': 'annual income',
-  'finance.retirement_age_band': 'retirement age',
-  'finance.existing_coverage': 'existing coverage',
-  'life_event.recent': 'recent life event',
-  'interests.tags': 'interests',
-  'residency.status': 'residency',
-};
-
-/** Taxonomy values are always {v: …} plus per-key extras — render them flat. */
-function factText(value) {
-  if (!value || typeof value !== 'object') return '—';
-  const v = value.v;
-  let body;
-  if (Array.isArray(v)) {
-    body = v.length
-      ? v.map((x) => (x && typeof x === 'object' ? Object.values(x).filter(Boolean).join(' ') : String(x))).join(', ')
-      : 'none';
-  } else if (typeof v === 'boolean') {
-    body = v ? 'yes' : 'no';
-  } else {
-    body = String(v ?? '—');
-  }
-  if (value.when) body += ` (${value.when})`;
-  // complete:false means "at least these" — say so rather than implying a
-  // closed list the ledger never claimed.
-  if (Array.isArray(v) && value.complete === false) body += ' — partial';
-  return body;
-}
-
-function ScoreDial({ label, value, hint }) {
-  const known = value != null;
-  return (
-    <div style={{ flex: 1, minWidth: 96 }}>
-      <div className="av2-microcaps" style={{ color: 'var(--ink-3)' }}>{label}</div>
-      <div
-        style={{ display: 'flex', alignItems: 'baseline', gap: 6, marginTop: 2 }}
-        title={known ? `${label} ${value}/100` : hint}
-      >
-        <span style={{ fontFamily: 'var(--font-mono)', fontSize: 26, fontWeight: 800, lineHeight: 1, color: known ? 'var(--ink)' : 'var(--ink-3)' }}>
-          {known ? value : '—'}
-        </span>
-        {known && <span style={{ fontFamily: 'var(--font-mono)', fontSize: 11, color: 'var(--ink-3)' }}>/100</span>}
-      </div>
-      {!known && <div style={{ fontSize: 11, color: 'var(--ink-3)', marginTop: 3 }}>{hint}</div>}
-    </div>
-  );
-}
-
-function ComponentRow({ name, c }) {
-  const assessed = c.state === 'assessed';
-  const max = Number(c.maxPoints) || 0;
-  // A penalty component's magnitude is what fills the bar; its sign shows in
-  // the number. Unknown draws no bar at all — an empty bar would read as zero.
-  const pct = assessed && max !== 0 ? Math.min(100, Math.abs(c.points / max) * 100) : 0;
-  const penalty = max < 0;
-  const label = COMPONENT_LABELS[name] || name.replace(/_/g, ' ');
-  // The reason moved ONTO the label (dotted = there is more here). It used to
-  // ride in a trailing column that ellipsed nearly every row — "reachable via
-  // marketi…", "no recent life event…" — so the one part that explains the
-  // number was the one part you could not read. Nothing is lost: the full text
-  // is one hover away, and the space it freed goes to the bar, which is what
-  // makes two components comparable at a glance.
-  return (
-    <div style={{ display: 'flex', alignItems: 'baseline', gap: 10, padding: '3px 0', fontSize: 12 }}>
-      <span style={{ width: 116, flex: 'none', color: assessed ? 'var(--ink-2)' : 'var(--ink-3)' }}>
-        {c.note
-          ? <HoverHint label={label} hint={c.note} ariaLabel={`${label}: ${c.note}`} />
-          : label}
-      </span>
-      <span style={{ width: 58, flex: 'none', fontFamily: 'var(--font-mono)', fontSize: 11, textAlign: 'right', color: assessed ? 'var(--ink)' : 'var(--ink-3)' }}>
-        {assessed ? `${c.points}` : '—'}<span style={{ color: 'var(--ink-3)' }}>/{max}</span>
-      </span>
-      <span aria-hidden="true" style={{ flex: 1, minWidth: 46, height: 4, borderRadius: 3, background: 'var(--surface-2)', overflow: 'hidden' }}>
-        {assessed && pct > 0 && (
-          <span style={{ display: 'block', width: `${pct}%`, height: '100%', background: penalty ? 'var(--bad)' : 'var(--accent-text)' }} />
-        )}
-      </span>
-    </div>
-  );
-}
-
-/**
- * The parts, grouped the way the config groups them. Shared by both grains
- * BECAUSE the two breakdowns are the same shape — the person's is literally a
- * copy of their winning lead's (`projectPersonScore`), so a reader who learns
- * to read one has learned to read the other.
- */
-function ScoreComponents({ breakdown }) {
-  const comps = breakdown?.components || {};
-  const groups = breakdown?.groups || {};
-  const completeness = breakdown?.completeness;
-  const meetNames = groups.meet?.components || [];
-  const buyNames = groups.buy?.components || [];
-  return (
-    <div style={{ padding: '4px 18px 12px' }}>
-      {meetNames.length > 0 && (
-        <>
-          <div className="av2-microcaps" style={{ padding: '6px 0 2px' }}>Reachability</div>
-          {meetNames.map((n) => comps[n] && <ComponentRow key={n} name={n} c={comps[n]} />)}
-        </>
-      )}
-      {buyNames.length > 0 && (
-        <>
-          <div className="av2-microcaps" style={{ padding: '10px 0 2px' }}>Potential</div>
-          {buyNames.map((n) => comps[n] && <ComponentRow key={n} name={n} c={comps[n]} />)}
-        </>
-      )}
-      {completeness && (
-        <div style={{ fontSize: 11.5, color: 'var(--ink-3)', paddingTop: 10 }}>
-          {completeness.assessed} of {completeness.total} components assessed
-          {completeness.assessed < completeness.total && ' — the unknowns above are questions nobody has been asked'}
-        </div>
-      )}
-    </div>
-  );
-}
-
-/**
- * THIS campaign's score for THIS person (per-campaign-lead-scoring.md §4).
- *
- * The scoring card on the person profile shows the WINNING lead's numbers — the
- * projection. For anyone with a single signup those are the same number, but a
- * person with two campaigns has two scores and only one of them surfaces there.
- * The drill-in is the page about one lead, so it is where that lead's own score
- * belongs.
- *
- * NULL is not zero. A lead the sweep has not reached yet says so in words
- * rather than showing a 0 nobody computed.
- */
-function LeadScoreBadge({ signup }) {
-  if (!signup) return null;
-  const scored = signup.score !== null && signup.score !== undefined;
-  if (!scored) {
-    return (
-      <span
-        className="av2-mono"
-        style={{ fontSize: 10.5, color: 'var(--ink-3)', flex: 'none' }}
-        title="The nightly sweep has not reached this lead yet."
-      >
-        NOT SCORED YET
-      </span>
-    );
-  }
-  return (
-    <span
-      style={{ display: 'inline-flex', alignItems: 'baseline', gap: 5, flex: 'none' }}
-      title={signup.scoredAt ? `Scored ${fmtDateTime(signup.scoredAt)}` : 'Lead score'}
-    >
-      <span className="av2-microcaps" style={{ color: 'var(--ink-3)' }}>lead score</span>
-      <span
-        className="av2-mono"
-        style={{ fontSize: 15, fontWeight: 800, color: 'var(--ink)', fontVariantNumeric: 'tabular-nums' }}
-      >
-        {signup.score}
-      </span>
-    </span>
-  );
-}
-
-/**
- * What actually happened, and when (per-campaign-lead-scoring.md §6).
- *
- * The components above say how much each thing is worth right now; this says
- * what the "thing" was. Weights are shown UNDECAYED — the score already has
- * the decay baked in as of `scoredAt`, so the gap between an event's full
- * weight and its contribution IS the age shown beside it. Rendering a decayed
- * figure here would just repeat the score twice and explain nothing.
- *
- * Absent for a lead nobody has messaged or called, which is most of them —
- * an empty section would read as "no response" rather than "no contact".
- */
-function ScoreEvents({ events, scoredAt }) {
-  const rows = Array.isArray(events) ? events : [];
-  if (!rows.length) return null;
-
-  const label = (e) => {
-    if (e.type === 'wa_read') return 'Read a WhatsApp from this campaign';
-    if (e.type === 'screening') {
-      const bits = [
-        e.verdict === 'qualified' ? 'Screening call: qualified'
-          : e.verdict === 'not_qualified' ? 'Screening call: not qualified'
-            : 'Screening call',
-        e.interest && `interest ${e.interest}`,
-        e.sentiment && `sentiment ${e.sentiment}`,
-        e.agreedToMeet === true && 'agreed to meet',
-        e.agreedToMeet === false && 'declined to meet',
-      ].filter(Boolean);
-      return bits.join(' · ');
-    }
-    return e.type;
-  };
-
-  return (
-    <Disclosure
-      label="Response events"
-      count={`${rows.length} EVENT${rows.length === 1 ? '' : 'S'}`}
-      indent={36}
-    >
-      {rows.map((e, i) => (
-        <div
-          key={`${e.type}-${e.at || i}`}
-          style={{ display: 'flex', alignItems: 'baseline', gap: 8, padding: '3px 0', fontSize: 12 }}
-        >
-          <span style={{ flex: 1, color: 'var(--ink-2)' }}>{label(e)}</span>
-          <span className="av2-mono" style={{ fontSize: 10.5, color: 'var(--ink-3)', flex: 'none' }}>
-            {e.at ? fmtDateTime(e.at) : 'undated'}
-            {Number.isFinite(e.ageDays) && ` · ${e.ageDays}d`}
-          </span>
-          <span
-            className="av2-mono"
-            style={{ width: 44, textAlign: 'right', flex: 'none', fontSize: 10.5, color: 'var(--ink-3)' }}
-            title="Full weight before decay"
-          >
-            {e.undecayedWeight != null ? `${e.undecayedWeight > 0 ? '+' : ''}${e.undecayedWeight}` : '—'}
-          </span>
-        </div>
-      ))}
-      <div style={{ fontSize: 11, color: 'var(--ink-3)', paddingTop: 6 }}>
-        Weights are shown at full strength; the score above applies decay as of{' '}
-        {scoredAt ? fmtDateTime(scoredAt) : 'the last rescore'}.
-      </div>
-    </Disclosure>
-  );
-}
-
-function EnrichmentCard({ enrichment }) {
-  const bd = enrichment?.breakdown;
-  const facts = enrichment?.facts || [];
-
-  return (
-    <Card
-      title="Scoring"
-      // The caption stamps come from the SOURCE lead (campaign-scoring-editor
-      // §4.4). When a breakdown survives but its source signup is gone —
-      // deleted, or a projection older than the pointer column — there is no
-      // honest stamp to show, and borrowing the person-pass one would caption
-      // this breakdown with an unrelated config.
-      meta={enrichment?.stampsUnavailable
-        ? 'SCORE SOURCE SIGNUP UNAVAILABLE'
-        : enrichment?.configVersion != null ? `CONFIG v${enrichment.configVersion}` : undefined}
-    >
-      <div style={{ padding: '12px 18px 6px', display: 'flex', gap: 18 }}>
-        <ScoreDial label="Meet" value={enrichment?.meetScore} hint="no signal yet" />
-        <ScoreDial
-          label="Buy"
-          value={enrichment?.buyScore}
-          hint="no facts to judge"
-        />
-      </div>
-
-      {/* A person does not have one worth — a fresh grad is a strong recruit
-          and a weak policyholder. These numbers are their BEST lead's (§4's
-          projection), so naming that lead is the difference between a number a
-          reader can act on and one they will misread the moment two campaigns
-          score the same person differently. */}
-      {enrichment?.scoreSource?.campaignName && (
-        <div style={{ padding: '0 18px 8px', fontSize: 11.5, color: 'var(--ink-3)' }}>
-          their best campaign —{' '}
-          <span style={{ color: 'var(--ink-2)', fontWeight: 600 }}>
-            {enrichment.scoreSource.campaignName}
-          </span>
-        </div>
-      )}
-
-      <ScoreComponents breakdown={bd} />
-
-      <ScoreEvents events={bd?.events} scoredAt={enrichment?.scoredAt} />
-
-      {facts.length > 0 && (
-        <Disclosure label="Fact ledger" count={`${facts.length} FACT${facts.length === 1 ? '' : 'S'}`} indent={36}>
-          {facts.map((f) => (
-            <div key={f.key} style={{ display: 'flex', gap: 10, fontSize: 12, padding: '3px 0', alignItems: 'baseline' }}>
-              <span style={{ width: 116, flex: 'none', color: 'var(--ink-3)' }}>{FACT_LABELS[f.key] || f.key}</span>
-              <span style={{ flex: 1, minWidth: 0, color: 'var(--ink)', fontWeight: 600, overflowWrap: 'anywhere' }}>
-                {factText(f.value)}
-              </span>
-              <span
-                className="av2-mono"
-                style={{ flex: 'none', fontSize: 10, color: 'var(--ink-3)' }}
-                title={`${f.source} · confidence ${f.confidence}${f.observedAt ? ` · ${fmtDate(f.observedAt)}` : ''}`}
-              >
-                {f.source}
-              </span>
-            </div>
-          ))}
-        </Disclosure>
-      )}
-
-      <ScoringMechanicsNote edition={enrichment?.stampsUnavailable ? null : enrichment?.configVersion} />
-    </Card>
-  );
-}
-
-/**
- * THIS CAMPAIGN'S customised sheet, on the lead (the follow-up ask: "no i
- * mean, the customised scoring metrics for this particular campaign").
- *
- * Shows the EXACT EDITION the stamp names — fetched lazily on first open,
- * never the currently-resolved sheet: the campaign may have moved to a newer
- * edition since this lead was scored, and captioning old points with new
- * rules is the §4.4 lie all over again. Weights render beside their house
- * values with a ↑/↓ marker on every deviation; the age curve and target
- * segments come from the same document. Effective weights fall back to the
- * breakdown's own maxPoints when an older edition's document omits a key.
- */
-function CampaignSheetPeek({ edition, breakdown, campaignId }) {
-  const [state, setState] = useState({ status: 'idle', sheet: null });
-  if (edition == null) return null;
-
-  const load = async (e) => {
-    if (!e.currentTarget.open || state.status === 'ready' || state.status === 'loading') return;
-    setState({ status: 'loading', sheet: null });
-    try {
-      setState({ status: 'ready', sheet: await fetchScoringEdition(edition) });
-    } catch {
-      setState({ status: 'error', sheet: null });
-    }
-  };
-
-  const sheet = state.sheet;
-  const cfg = sheet?.configJson || {};
-  const house = sheet?.houseDefault || {};
-  const comps = breakdown?.components || {};
-  // Effective weight: the edition's document, else the breakdown's own max —
-  // which IS the weight that actually applied.
-  const weightFor = (name, leadGrain) => {
-    const map = leadGrain ? cfg.leadComponents : cfg.components;
-    return map?.[name]?.maxPoints ?? comps[name]?.maxPoints ?? null;
-  };
-  const houseFor = (name, leadGrain) => {
-    const map = leadGrain ? house.leadComponents : house.components;
-    return map?.[name]?.maxPoints ?? null;
-  };
-
-  // "peaks 30-44" from the curve: the segments at the curve's own maximum.
-  const curve = Array.isArray(cfg.ageCurve) ? cfg.ageCurve : [];
-  let peak = null;
-  if (curve.length) {
-    const top = Math.max(...curve.map((s) => s.value));
-    const ranges = [];
-    let lo = 0;
-    for (const seg of curve) {
-      const hi = seg.upTo == null ? null : seg.upTo;
-      if (seg.value === top) ranges.push(hi == null ? `${lo}+` : `${lo}–${hi}`);
-      lo = hi == null ? lo : hi + 1;
-    }
-    peak = ranges.join(', ');
-  }
-  const segments = Array.isArray(cfg.targetSegments) ? cfg.targetSegments : [];
-
-  const NAMES = [
-    ['meet', 'Reachability', [['engagement', false], ['contactability', false], ['market_fit', false], ['response', true], ['screening', true]]],
-    ['buy', 'Potential', [['life_events', false], ['family_gap', false], ['capacity', false], ['coverage_headroom', false], ['age', false]]],
-  ];
-
-  return (
-    <details style={{ borderTop: '1px solid var(--line)' }} onToggle={load}>
-      <summary style={{ display: 'flex', alignItems: 'center', gap: 8, padding: '9px 16px', cursor: 'pointer', fontSize: 11.5, fontWeight: 700, color: 'var(--ink-2)' }}>
-        <span className="disc-caret" aria-hidden="true" style={{ color: 'var(--ink-3)', display: 'inline-block', transition: 'transform .12s ease' }}>▸</span>
-        This campaign's scoring sheet
-        <span style={{ flex: 1 }} />
-        <span style={{ fontFamily: 'var(--font-mono)', fontSize: 10, color: 'var(--ink-3)' }}>EDITION #{edition}</span>
-      </summary>
-      <div style={{ padding: '2px 16px 12px 36px' }}>
-        {state.status === 'loading' && <div style={{ fontSize: 12, color: 'var(--ink-3)', padding: '4px 0' }}>Loading the edition…</div>}
-        {state.status === 'error' && (
-          <div style={{ fontSize: 12, color: 'var(--ink-3)', padding: '4px 0' }}>
-            Couldn't load edition #{edition} — the weights in the breakdown above (the “/N” figures) are the ones that applied.
-          </div>
-        )}
-        {state.status === 'ready' && (
-          <>
-            {NAMES.map(([key, title, names]) => (
-              <div key={key}>
-                <div className="av2-microcaps" style={{ padding: '6px 0 2px' }}>{title}</div>
-                {names.map(([name, leadGrain]) => {
-                  const w = weightFor(name, leadGrain);
-                  const h = houseFor(name, leadGrain);
-                  const differs = w != null && h != null && w !== h;
-                  return (
-                    <div key={name} style={{ display: 'flex', alignItems: 'baseline', gap: 10, padding: '2px 0', fontSize: 12 }}>
-                      <span style={{ width: 148, flex: 'none', color: 'var(--ink-2)' }}>{COMPONENT_LABELS[name] || name}</span>
-                      <span className="av2-mono" style={{ width: 34, textAlign: 'right', flex: 'none', fontWeight: differs ? 800 : 500, color: differs ? 'var(--ink)' : 'var(--ink-2)' }}>
-                        {w ?? '—'}
-                      </span>
-                      <span className="av2-mono" style={{ fontSize: 10.5, color: 'var(--ink-3)' }}>
-                        house {h ?? '—'}
-                        {differs && (
-                          <span style={{ color: w > h ? 'var(--accent-text)' : 'var(--warn)', fontWeight: 700 }}>
-                            {' '}{w > h ? '↑' : '↓'} {w > h ? '+' : ''}{Math.round((w - h) * 10) / 10}
-                          </span>
-                        )}
-                      </span>
-                    </div>
-                  );
-                })}
-              </div>
-            ))}
-            <div style={{ fontSize: 12, color: 'var(--ink-2)', paddingTop: 8 }}>
-              <span className="av2-microcaps" style={{ display: 'block', paddingBottom: 2 }}>Age dial</span>
-              {peak ? <>full weight at ages <span style={{ fontWeight: 700, color: 'var(--ink)' }}>{peak}</span>, ramping down outside</> : 'house curve'}
-            </div>
-            <div style={{ fontSize: 12, color: 'var(--ink-2)', paddingTop: 8 }}>
-              <span className="av2-microcaps" style={{ display: 'block', paddingBottom: 2 }}>Target market</span>
-              {segments.length
-                ? segments.map((s, i) => (
-                  <span key={i}>
-                    {i > 0 && ' · '}
-                    {[s.language && (SEGMENT_LANGUAGE_LABELS[s.language] || s.language), s.ethnicity].filter(Boolean).join(' / ')}
-                    {typeof s.weight === 'number' && s.weight !== 1 ? ` ×${s.weight}` : ''}
-                  </span>
-                ))
-                : 'everyone equally'}
-            </div>
-            {campaignId && (
-              <div style={{ paddingTop: 10 }}>
-                <Link to={`/admin/campaigns/${campaignId}`} style={{ fontSize: 11.5, fontWeight: 700 }}>
-                  Tune this sheet on the campaign page →
-                </Link>
-              </div>
-            )}
-          </>
-        )}
-      </div>
-    </details>
-  );
-}
-
-const SEGMENT_LANGUAGE_LABELS = { en: 'English', zh: 'Mandarin', ms: 'Malay', ta: 'Tamil' };
-
-/**
- * The MECHANICS, on the page (asked for directly: "can you put somewhere,
- * the scoring mechanics on the lead?"). The same plain language the console
- * uses everywhere, tucked behind a disclosure so it teaches without
- * shouting. Shared by BOTH score cards — the person card renders a copy of a
- * lead's breakdown, so one explanation is the truth for both.
- *
- * Wordings are deliberately DISTINCT from the cards' own microcopy (the sum
- * line, the completeness line) so nothing on screen repeats itself verbatim.
- */
-function ScoringMechanicsNote({ edition }) {
-  const p = { margin: 0 };
-  const b = (text) => <span style={{ fontWeight: 700, color: 'var(--ink)' }}>{text}</span>;
-  return (
-    <Disclosure label="How this score works" indent={36}>
-      <div style={{ fontSize: 12, color: 'var(--ink-2)', lineHeight: 1.6, display: 'grid', gap: 8, maxWidth: 660, paddingTop: 2 }}>
-        <p style={p}>
-          {b('A points checklist, not a grade.')} Ten questions we would ask about any
-          lead, each worth a set number of points. Evidence earns points; a question
-          never asked shows “—” and earns nothing — {b('unknown is never treated as low')}.
-        </p>
-        <p style={p}>
-          {b('Two kinds of evidence.')} What the person {b('did')} — signed up, verified
-          their phone, read our WhatsApp, took the screening call — and what we
-          {' '}{b('know')} about them — age, children, income — from form answers and
-          call transcripts. Each row’s dotted label explains its own points on hover.
-        </p>
-        <p style={p}>
-          {b('The three numbers.')} The headline is every earned point added together
-          (capped at 100). Meet is the reachability rows against their own maximum —
-          “will they take a consultant’s call”. Buy is the potential rows against
-          theirs — “would they actually buy”. Buy stays “—” until at least one real
-          fact is known, so ignorance can never fake a number.
-        </p>
-        <p style={p}>
-          {b('Old signals fade.')} Activity counts for about half after six months,
-          life events after a year. The score is recomputed nightly and when
-          something new happens; “scored” above is its as-of stamp.
-        </p>
-        <p style={p}>
-          {b('The rulebook is per campaign.')} These weights come from
-          {edition != null ? <> this campaign’s scoring sheet — edition #{edition}</> : ' the campaign’s scoring sheet'},
-          tunable from the campaign page (draft → preview on real leads → approve;
-          every edition is kept, with who approved it).
-        </p>
-      </div>
-    </Disclosure>
-  );
-}
-
-/**
- * WHY this lead scored what it scored — the drill-in's own breakdown.
- *
- * The hero line states the number; this states the working, in the same parts
- * and the same order as the person card, because the two breakdowns are the
- * same shape.
- *
- * It reads the PROSPECT ROW, which is the only per-lead source on this page:
- * `journey.enrichment` is the person's projection of their best lead (§4), so
- * on anyone's second campaign it would explain a number this page is not
- * showing. `?include=profile` already serializes the whole prospect model, so
- * the breakdown, both halves and the config stamp arrive with no extra request.
- *
- * Absent until the sweep has scored this lead — `scoredConfigVersion` is the
- * stamp that means "scored", and an empty card would imply a verdict nobody
- * has reached.
- */
-function LeadScoreCard({ prospect, campaignName, campaignId, hasPersonCard }) {
-  const bd = prospect?.scoreBreakdown || null;
-  if (!bd || prospect.scoredConfigVersion == null) return null;
-  const stamp = [
-    `CONFIG v${prospect.scoredConfigVersion}`,
-    prospect.scoreComputedAt ? `SCORED ${fmtDateTime(prospect.scoreComputedAt).toUpperCase()}` : null,
-  ].filter(Boolean).join(' · ');
-  return (
-    <Card title="Lead score" meta={stamp}>
-      <div style={{ padding: '12px 18px 6px', display: 'flex', gap: 18 }}>
-        <ScoreDial label="Meet" value={prospect.meetScore} hint="no signal yet" />
-        <ScoreDial label="Buy" value={prospect.buyScore} hint="no facts to judge" />
-      </div>
-
-      {/* The total is NOT an average of the two halves — average 50 and 16 and
-          you get 33, not the 48 in the hero. It is the points below, added up
-          and capped. Stated here because this is the one place the parts and
-          the whole are on screen together, which is exactly where a reader
-          tries the arithmetic and concludes the page is lying. */}
-      {prospect.score != null && (
-        <div style={{ padding: '0 18px 4px', fontSize: 11.5, color: 'var(--ink-3)' }}>
-          <span className="av2-mono" style={{ color: 'var(--ink-2)', fontWeight: 700 }}>{prospect.score}</span>
-          {' = the points below, added up and capped at 100 — not an average of the two halves.'}
-        </div>
-      )}
-
-      {/* §4: the profile card shows this person's BEST lead. On a second
-          campaign that is a different lead with different numbers, so the two
-          cards disagreeing is the system working — say so before a reader
-          reads it as a bug. But only when that card exists: a consumer-less
-          (or never-person-scored) lead has no profile scoring card, and
-          pointing at one would send the reader hunting for a surface that
-          isn't there. */}
-      <div style={{ padding: '0 18px 8px', fontSize: 11.5, color: 'var(--ink-3)' }}>
-        {campaignName
-          ? <><span style={{ color: 'var(--ink-2)', fontWeight: 600 }}>{campaignName}</span>{' only'}</>
-          : 'This campaign only'}
-        {hasPersonCard ? ' — the profile card scores their best campaign.' : '.'}
-      </div>
-
-      <ScoreComponents breakdown={bd} />
-
-      <ScoreEvents events={bd.events} scoredAt={prospect.scoreComputedAt} />
-
-      <CampaignSheetPeek
-        edition={prospect.scoredConfigVersion}
-        breakdown={bd}
-        campaignId={campaignId}
-      />
-
-      <ScoringMechanicsNote edition={prospect.scoredConfigVersion} />
-    </Card>
-  );
-}
-
-function Disclosure({ label, count, children, indent = 34 }) {
-  return (
-    <details style={{ borderTop: '1px solid var(--line)' }}>
-      <summary style={{ display: 'flex', alignItems: 'center', gap: 8, padding: '9px 16px', cursor: 'pointer', fontSize: 11.5, fontWeight: 700, color: 'var(--ink-2)' }}>
-        <span className="disc-caret" aria-hidden="true" style={{ color: 'var(--ink-3)', display: 'inline-block', transition: 'transform .12s ease' }}>▸</span>
-        {label}
-        <span style={{ flex: 1 }} />
-        {count && <span style={{ fontFamily: 'var(--font-mono)', fontSize: 10, color: 'var(--ink-3)' }}>{count}</span>}
-      </summary>
-      <div style={{ padding: `2px 16px 12px ${indent}px` }}>{children}</div>
-    </details>
-  );
-}
-
-function GlyphTile({ family, size = 22 }) {
-  const t = FAMILY_TILE[family] || FAMILY_TILE.generic;
-  return (
-    <span aria-hidden="true" style={{ width: size, height: size, flex: 'none', borderRadius: size > 24 ? 9 : 7, background: t.bg, color: t.fg, display: 'flex', alignItems: 'center', justifyContent: 'center', fontSize: size > 24 ? 13 : 11 }}>
-      {t.glyph}
-    </span>
-  );
-}
 
 // ── the page ────────────────────────────────────────────────────────────────
